Update aznsent.py
I added a block of code after TextBlob Analysis.  This code was working on a subset of 100 reviews.  It will take a long time if you do the whole dataset.  I'm sure there is a way to wrap all this up into a function but I don't know how to do it.  The Lemmatizer is a little suspect, I'm not sure that it did anything.  I think it supposed to take of 'ing', 'ed', 'en', and the like but I can tell that it did it.  My code adds a few more columns to the dataframe so it might break the elastic search import if its looking for a specific number of columns.  The removal of 10 most common and 10 rarest is arbitrary.  You can removed the 5 most common or the 30 rarest.  Whatever fits this analysis.